Update HTML, CSS and JavaScript
Updated the app to display a 'words per minute' value when the test finishes, and a reset button that can be toggled to.

Calculating the WPM has been left to a separate function that iterates through the words array at the end, calculating results based on the classes added to each letter during the test, rather than incrementing the values inside the character checks.